test: split big TestCase class
